+++ Academy.DataAccess/Mapping/TelephoneMap.cs
@@ -15,7 +15,7 @@ namespace Academy.DataAccess.Mapping
         {
             ToTable("Telephone");
             HasKey(t => t.TelephoneId);
-            Property(t => t.TelephoneId).IsRequired().HasColumnName("TelephoneId").HasDatabaseGeneratedOption(DatabaseGeneratedOption.None).HasColumnType("BIGINT");
+            Property(t => t.TelephoneId).IsRequired().HasColumnName("TelephoneId").HasDatabaseGeneratedOption(DatabaseGeneratedOption.None).HasColumnType("INT");
             Property(t => t.AreaCode).IsRequired().HasColumnName("AreaCode").HasMaxLength(3).HasColumnType("NVARCHAR");
             Property(t => t.Exchange).IsRequired().HasColumnName("Exchange").HasMaxLength(3).HasColumnType("NVARCHAR");
             Property(t => t.PhoneNumber).IsRequired().HasColumnName("PhoneNumber").HasMaxLength(4).HasColumnType("NVARCHAR");
